Minor fixes 7th chapter

+++ src/main/java/lab/dao/jdbc/CountryDao.java
@@ -33,7 +33,7 @@ public class CountryDao extends NamedParameterJdbcDaoSupport {
 
     private static final RowMapper<Country> COUNTRY_ROW_MAPPER = (resultSet, i) ->
             new SimpleCountry(
-                    resultSet.getInt("id"),
+                    resultSet.getLong("id"),
                     resultSet.getString("name"),
                     resultSet.getString("code_name"));
